Updated the project so the include/gfesdk is kept empty in the repo.
Added explanations in the README.md file to explain how to compile.

Added guards, a pre-build script will run and setup Program.hpp and either set or not, a define flag to tell the rest of the program if the GfeSDK is available or not.

This way if we compile and GfeSDK is not available, we instead compile a simple error message explaining we are missing the GfeSDK, otherwise we compile the program as expected.